+++ src/solana/solanaInfo.ts
@@ -1,8 +1,60 @@
-import { EdgeCurrencyInfo } from 'edge-core-js/types'
+import { EdgeCurrencyInfo, EdgeTokenMap } from 'edge-core-js/types'
 
 import { makeOuterPlugin } from '../common/innerPlugin'
+import { makeMetaTokens } from '../common/tokenHelpers'
 import type { SolanaTools } from './SolanaTools'
 import type { SolanaNetworkInfo } from './solanaTypes'
+
+const builtinTokens: EdgeTokenMap = {
+  kshrEkxuc7zPAvLxvabxoERKxK6BfariPcjBoiHvM7B: {
+    currencyCode: 'GMT',
+    displayName: 'STEPN',
+    denominations: [{ name: 'GMT', multiplier: '1000000000' }],
+    networkLocation: {
+      contractAddress: 'kshrEkxuc7zPAvLxvabxoERKxK6BfariPcjBoiHvM7B'
+    }
+  },
+  CWE8jPTUYhdCTZYWPTe1o5DFqfdjzWKc9WKz6rSjQUdG: {
+    currencyCode: 'LINK',
+    displayName: 'Wrapped Chainlink',
+    denominations: [{ name: 'LINK', multiplier: '1000000' }],
+    networkLocation: {
+      contractAddress: 'CWE8jPTUYhdCTZYWPTe1o5DFqfdjzWKc9WKz6rSjQUdG'
+    }
+  },
+  HZ1JovNiVvGrGNiiYvEozEVgZ58xaU3RKwX8eACQBCt3: {
+    currencyCode: 'PYTH',
+    displayName: 'Pyth Network',
+    denominations: [{ name: 'PYTH', multiplier: '1000000' }],
+    networkLocation: {
+      contractAddress: 'HZ1JovNiVvGrGNiiYvEozEVgZ58xaU3RKwX8eACQBCt3'
+    }
+  },
+  rndrizKT3MK1iimdxRdWabcF7Zg7AR5T4nud4EkHBof: {
+    currencyCode: 'RENDER',
+    displayName: 'Render Token',
+    denominations: [{ name: 'RENDER', multiplier: '100000000' }],
+    networkLocation: {
+      contractAddress: 'rndrizKT3MK1iimdxRdWabcF7Zg7AR5T4nud4EkHBof'
+    }
+  },
+  EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v: {
+    currencyCode: 'USDC',
+    displayName: 'USD Coin',
+    denominations: [{ name: 'USDC', multiplier: '1000000' }],
+    networkLocation: {
+      contractAddress: 'EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v'
+    }
+  },
+  Es9vMFrzaCERmJfrF4H2FYD4KCoNkY11McCe8BenwNYB: {
+    currencyCode: 'USDT',
+    displayName: 'Tether',
+    denominations: [{ name: 'USDT', multiplier: '1000000' }],
+    networkLocation: {
+      contractAddress: 'Es9vMFrzaCERmJfrF4H2FYD4KCoNkY11McCe8BenwNYB'
+    }
+  }
+}
 
 const networkInfo: SolanaNetworkInfo = {
   rpcNodes: [
@@ -42,12 +94,13 @@ const currencyInfo: EdgeCurrencyInfo = {
   // Deprecated:
   defaultSettings: {},
   memoType: 'text',
-  metaTokens: []
+  metaTokens: makeMetaTokens(builtinTokens)
 }
 
 export const solana = makeOuterPlugin<SolanaNetworkInfo, SolanaTools>({
   currencyInfo,
   networkInfo,
+  builtinTokens,
 
   checkEnvironment: () => {
     if (global.BigInt == null) {